pyproject.toml
[project]
name = "django-ninja-template"
version = "1.0.0"
description = "基于 Django + django-ninja 的 RESTful API 后端模板项目"
requires-python = ">=3.12"
dependencies = [
    "Django>=4.2,<5.0",
    "django-ninja>=1.6.2",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-django>=4.8.0",
    "ruff>=0.4.0",
    "mypy>=1.9.0",
    "django-stubs>=4.2.0",
    "djangorestframework-stubs>=3.14.0",
]

[tool.ruff]
target-version = "py312"
line-length = 88
select = [
    "E",     # pycodestyle errors
    "W",     # pycodestyle warnings
    "F",     # pyflakes
    "I",     # isort
    "B",     # flake8-bugbear
    "C4",    # flake8-comprehensions
    "UP",    # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.12"
plugins = ["mypy_django_plugin.main", "mypy_django_stubs.plugin"]
strict = true
disallow_untyped_defs = true
disallow_any_generics = true
warn_return_any = true
warn_unused_ignores = true

[tool.django-stubs]
django_settings_module = "config.settings"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
python_files = ["test_*.py", "*_test.py"]
addopts = "-v --tb=short"
testpaths = ["tests"]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
